توثيق الحساب وكلمات المرور

+++ back/app/Http/Controllers/DashboardController.php
@@ -141,7 +141,7 @@ class DashboardController extends Controller
         }
     }
 
-    private function calculateGrowthPercentage(int $current, int $previous): float
+    private function calculateGrowthPercentage(float $current, float $previous): float
     {
         if ($previous === 0) {
             // إذا كانت الفترة السابقة صفر والحالية أكبر من صفر فالنمو 100%، وإلا فهو 0%
